Add ICU4J prebuilt used by ADT and layoutlib [DO NOT MERGE]
Add ICU v50.1.1 prebuilt. It is bundled with the SDK and used by
layoutlib to determine the character orientation (right to left or left
to right) of the locale with which the layout should be rendered. This
is the Java version corresponding to the C version in external/icu4c

Change-Id: Ib1b6cf7b5bcb3cd9ad2c103850dafa500f72a7dc
(cherry picked from commit 9d4363b02c38e89a11323631eadcd279cc342c30)