Update espresso prebuilts. DO NOT MERGE ANYWHERE
This updates espresso prebuilts only. An earlier CL (see ag/736015)
tried to update everything, but an android-support-test change caused
clockwork builds to break in Emerald.

This means that the espresso JARs might reference a different
version of android-support-test than what is present in this directory
but things seem to work with existing tests.

Change-Id: Ica4f2e89ab8040d7f9cfd6602065ac921e3ad953