Add a Robolectric prebuilt directory with build rules.
Robolectric and its dependencies are inside of the
prebuilts/maven_repo/robolectric directory, but are stuck in
nested directories. By flatten them into a single lib directory,
we can pass them to Robolectric as a dependency directory as a
system property at runtime. This allows Robolectric to run from
the command line and from inside of an Android make file.

With this, this patch also adds build targets to Android.mk which
build Robolectric and also a test target which runs specified
Robolectric tests.

Change-Id: Iba96495475292f50dfdd5116d3e1abdc97a679cf