Set min sdk of platform-robolectric-utils to o-b1
And transitively to platform-robolectric-prebuilt which is importing
it.
This is required because:
- platform-robolectric-utils.jar contains calls to
  MethodHandle.invoke / MethodHandle.invokeExact.
- platform-robolectric-prebuilt jack file is required as classpath
  entry of so check targets.
It should be OK since those 2 jack files are never used in a dex file,
they are just required for check targets, the real executed code is in
the class files.

Test: mmm prebuilts/misc/common/robolectric
Change-Id: I16d46d0a3dfd6cb4c567c6ff61564d3bdb0f6dfb
(cherry picked from commit 9684ec67c15b02828b8ce629396b64ea304f7dc8)