Remove TransportManager from Robolectric android jars
In order to be able to migrate to Robolectric 3.5.1, B&R tests need that
the android jar does NOT contain TransportManager otherwise Robolectric
picks it up instead of real implementation. As agreed with Robolectric
team, the android jar should not contain the TransportManager because
it's not available to apps and runs in the system process. The following
files were removed from the jars in this CL (if they were there in the
first place):
* com/android/server/backup/TransportManager$1.class
* com/android/server/backup/TransportManager.class
* com/android/server/backup/TransportManager$RebindOnTimeoutHandler.class
* com/android/server/backup/TransportManager$TransportBoundListener.class
* com/android/server/backup/TransportManager$TransportConnection.class
* com/android/server/backup/TransportManager$TransportReadyCallback.class

They were removed with: zip -d <jar_file> <class_files_above>

Also had to include asm-6.0.jar in the
platform-robolectric-3.5.1-prebuilt target to be able to write a custom
SandboxClassLoader (see other CL with this topic).

Test: m -j RunSettingsRoboTests RunFrameworksServicesRoboTests
RunSuwRoboTests RunTurboRoboTests RunCarSetupWizardRoboTests
RunCarSettingsRoboTests RunTvSettingsRoboTests
RunStorageManagerRoboTests RunSettingsLibRoboTests RunKeyChainRoboTests
RunBluetoothRoboTests RunCarMessengerRoboTests RunEmergencyInfoRoboTests
RunStorageManagerGoogleRoboTests RunCarSetupWizardLibRoboTests
RunSettingsGoogleRoboTests RunManagedProvisioningRoboTests
RunWearSignalDetectorRoboTests
Bug: 69254249
Change-Id: Ic86b1bf2b69f2da417c36d60cbb4d0a7a0490466